formatting improved, db mod added new imports
