+++ src/main.rs
@@ -1,16 +1,17 @@
 use anyhow::Result;
 use dotenv::dotenv;
-use state::State;
-use tmi::client::ConnectError;
 use tmi::Client;
+use tmi::client::ConnectError;
 use tokio::select;
 use tokio::signal::ctrl_c;
-
 use twitch_api2::{helix::channels::GetChannelInformationRequest, TwitchClient};
+
+use state::State;
 
 mod chatter;
 mod commands;
 mod content;
+mod db;
 mod duel;
 mod messaging;
 mod models;
